Sky is the limit, let's bring that limit higher
In this web stack debugging task, we are testing how well our web server setup featuring Nginx is doing under pressure ,

Apache Bench is used which is a quite popular tool in the industry. It allows you to simulate HTTP requests to a web server. In this case, I will make 2000 requests to my server with 100 requests at a time.